Add files via upload
AtliQ Grands, a well-known chain of luxury hotels in India, has a rich 20-year history in hospitality. However, they are currently facing challenges with decreasing market share and revenue in the luxury and business hotel segment due to competition and management decisions. To address this, the managing director is initiating a strategic move to integrate "Business and Data Intelligence" into their operations, aiming to regain market share and boost revenue.